# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "paths": {
    "/estacao/media/{mes}/{ano}/{codigo}": {
      "get": {
        "summary": "Get Monthly Avg",
        "operationId": "get_monthly_avg_estacao_media__mes___ano___codigo__get",
        "parameters": [
          {
            "name": "mes",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Mes"
            }
          },
          {
            "name": "ano",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ano"
            }
          },
          {
            "name": "codigo",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Codigo"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {

                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/estacao/media/{ano}/{codigo}": {
      "get": {
        "summary": "Get Annual Avg",
        "operationId": "get_annual_avg_estacao_media__ano___codigo__get",
        "parameters": [
          {
            "name": "ano",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Ano"
            }
          },
          {
            "name": "codigo",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Codigo"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {

                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/estacao/{mes}/{ano}/{codigo}": {
      "get": {
        "summary": "Get All Measurements",
        "operationId": "get_all_measurements_estacao__mes___ano___codigo__get",
        "parameters": [
          {
            "name": "mes",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Mes"
            }
          },
          {
            "name": "ano",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ano"
            }
          },
          {
            "name": "codigo",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Codigo"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {

                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/estacao/start": {
      "post": {
        "summary": "Manual Trigger",
        "operationId": "manual_trigger_estacao_start_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {

                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}